+++ enjoy/src/renderer/components/chats/index.ts
@@ -3,5 +3,7 @@ export * from "./chat-agents";
 export * from "./chat-agent-form";
 export * from "./chat-card";
 export * from "./chat-form";
-export * from "./chat-session";
+export * from "./chat-input";
+export * from "./chat-message";
+export * from "./chat-sessions";
 export * from "./chat-sidebar";
