+++ enjoy/src/renderer/context/index.ts
@@ -1,8 +1,9 @@
 export * from "./ai-settings-provider";
 export * from "./app-settings-provider";
 export * from "./chat-provider";
+export * from "./chat-session-provider";
 export * from "./course-provider";
 export * from "./db-provider";
-export * from './hotkeys-settings-provider'
+export * from "./hotkeys-settings-provider";
 export * from "./media-player-provider";
 export * from "./theme-provider";
